+++ Infrastructure/Repository/IRepositoryUser.cs
@@ -10,7 +10,7 @@ namespace Infrastructure.Repository
     public interface IRepositoryUser
     {
         IEnumerable<User> GetUsers();
-        User GetUserByID(int id);
+        User GetUserByID(long id);
         User GetUserByIDForLogin(long id);
         User Save(User user);
         User GetUsersForLogin(string email, string password);
